Move conditional logic into chart components
