+++ docs/docs/models/Benchmarks/Performance/index.mdx
@@ -8,15 +8,11 @@ pagination_prev: null
 full_width: true
 ---
 
-import BrowserOnly from '@docusaurus/BrowserOnly';
+import PerformanceChart from '@site/src/components/chart/performanceChart/performanceChart';
 
 # Performance Benchmarks
 
 Use this page to see performance benchmarks across all UpscalerJS models.
+
+<PerformanceChart databasePath="/assets/performance.sql"></PerformanceChart>
  
-<BrowserOnly>
-  {() => {
-    const { PerformanceChart } = require('@site/src/components/chart/performanceChart/performanceChart');
-    return (<PerformanceChart databasePath="/assets/performance.sql" />);
-  }}
-</BrowserOnly>
